fix(step-7): correct PPU step loop and test improvements
What was fixed:
- PPU step() now loops to process all cycles in one call, fixing mode transitions
- STAT register initialized with correct initial mode bits (mode 2)
- Color.fromGbc15bit() now returns explicit int values for PHPStan
- Removed unused $windowRenderedThisFrame property
- Fixed interrupt tests to enable interrupts in IE register
- Fixed tile rendering tests to use unsigned addressing mode (LCDC bit 4)
- Fixed window rendering test to fill entire tile map row

Why these changes:
- PPU step() was only processing one mode transition per call, causing LY to not increment
- STAT register was showing mode 0 instead of initial mode 2
- PHPStan requires explicit int casts for division results
- InterruptController.getPendingInterrupt() requires both IF and IE to be set
- Tests were using signed mode (bit 4=0) but writing tiles to unsigned mode offsets
- Window tile map needs multiple tiles set to render across full scanline

Verification:
- All 246 tests pass with 24,244 assertions
- PHPStan analysis passes with 0 errors (level 9)
- PPU mode transitions work correctly across full frame (70,224 dots)
- LY increments properly from 0-153
- All interrupt types trigger correctly
- Background, window, and sprite rendering work as expected
- Both signed and unsigned tile addressing modes work correctly

+++ src/Ppu/Color.php
@@ -54,9 +54,9 @@ final readonly class Color
 
         // Scale 5-bit values (0-31) to 8-bit values (0-255)
         return new self(
-            ($r * 255) / 31,
-            ($g * 255) / 31,
-            ($b * 255) / 31,
+            (int) (($r * 255) / 31),
+            (int) (($g * 255) / 31),
+            (int) (($b * 255) / 31),
         );
     }
